#19 implementing the dependency on lights - removed TSL sensor from outdoor sensors and added into light capable - updated support and outdoor and rooms, specifications and factories - new specification concept, older observers rewritten using this one - new specification to get satisfied either if we have light below   treshold or a control on (i.e. the light itself is on, then we let the   motion sensor switch it off) - test cases added - I2C sensor marked as not initialized if setup fails

+++ rpi/src/main/java/cz/pojd/rpi/sensors/i2c/TSL2561LightSensor.java
@@ -83,13 +83,13 @@ public class TSL2561LightSensor extends I2CSensor implements Sensor, DisposableB
      * Device lux range 0.1 - 40,000+ see https://learn.adafruit.com/tsl2561/overview
      */
     private double readLux() throws IOException {
-	int ambient = readFull();
-	int ir = readIR();
+	double ambient = readFull();
+	double ir = readIR();
 
 	if (ambient >= 0xffff || ir >= 0xffff) // value(s) exceed(s) datarange
 	    throw new IOException("Gain too high. Values exceed range.");
 
-	double ratio = (ir / (float) ambient);
+	double ratio = ir / ambient;
 
 	if (LOG.isDebugEnabled()) {
 	    LOG.debug("IR Result:" + ir);
